e2b/requirements.txt
e2b-code-interpreter==1.0.0
python-dotenv==1.0.1
streamlit==1.39.0
pandas==2.2.3
plotly==5.24.1
numpy==2.4.1
snowflake-connector-python==3.12.3
